requirements.txt
requests>=2.31.0
illumio>=1.1.3
python-dotenv>=1.0.0
